rename feature-section to card-component
